mcp-cep/pyproject.toml
[project]
name = "mcp-cep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.4.1",
    "python-dotenv>=1.0.1",
]
